feature (console): disable query and subscription root fields
## Description 🔖

Adding root field level section for permissions for select and subscriptions.
One can choose specific root permissions per role that can be enabled or disabled. This is done on a role basis.

By default the permissions does not exist. When this is the case, all permission are enabled per default.
Once a permission has been tampered with, the permissions work as expected by the ui.

## Related Issues 👾

Jira issue: https://hasurahq.atlassian.net/browse/CON-259

## Solution and Design 🎨

Feature showcase video recording: https://app.iterspace.com/recording/6faa1c87-fd17-450b-ac57-3a788bfcc6e4

The solution is made in our legacy part of the Data -> Permissions-tab. The section is going to be migrated in a short future.

### Steps to test

- go to Data tab -> any table -> Permissions tab
- create and edit roles by playing around with the new "Root fields permissions" section

## Known issues
The vertical alignment of the "Don't ask me again" is not the expected one.
It is due to the current `Checkbox` component that includes a label and an error message div, that is not needed in this use case.

As discussed with @nicoinch I could refactor the current `Checkbox` component into a `FormCheckbox` component, make a new atom component `Checkbox` with the input and right label only, and use it inside the `FormCheckbox`.

This will be done on another card ([CON-415](https://hasurahq.atlassian.net/browse/CON-415))

![image](https://user-images.githubusercontent.com/59067245/181784002-19b07ff2-5eb4-4867-92d8-4d2e51afbeb3.png)

### Typescript & State management

- [x] Migrate modified JS files to TS

### Pro console
- [x] Pro console build has been verified. If this check is not relevant, please remove this line.

## Changelog ✍️

- [x] `CHANGELOG.md` is updated with user-facing content relevant to this PR. If no changelog is required, then add the `no-changelog-required` label.

PR-URL: https://github.com/hasura/graphql-engine-mono/pull/4873
Co-authored-by: Rikin Kachhia <54616969+rikinsk@users.noreply.github.com>
Co-authored-by: Luca Restagno <59067245+lucarestagno@users.noreply.github.com>
GitOrigin-RevId: 51aa05af3c4de4dea957e7a12aaf4108e797d3ba